+++ appointment-scheduling-service/src/main/java/com/example/appointmentschedulingservice/service/AppointmentSchedulingService.java
@@ -1,22 +1,20 @@
 package com.example.appointmentschedulingservice.service;
 
 
-
 import com.api.project2.model.generated.appointmentscheduling.*;
-
-import java.math.BigDecimal;
 
 
 import java.util.Optional;
 
 public interface AppointmentSchedulingService {
 
-    Optional<AppointmentDetails> getAppointmentById(BigDecimal appointmentId);
+    Optional<AppointmentDetails> getAppointmentById(Integer appointmentId);
 
     CreateAppointmentResponse createAppointment(CreateAppointmentRequest createAppointmentRequest);
 
-    UpdateAppointmentResponse updateAppointment(UpdateAppointmentRequest updateAppointmentRequest, BigDecimal appointmentId);
+    UpdateAppointmentResponse updateAppointment(UpdateAppointmentRequest updateAppointmentRequest, Integer appointmentId);
 
-    DeleteAppointmentResponse deleteAppointment(BigDecimal appointmentId);
+    DeleteAppointmentResponse deleteAppointment(Integer appointmentId);
+
 
 }
